detr/requirements.txt
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.7
fonttools==4.47.0
importlib-resources==6.1.1
kiwisolver==1.4.5
matplotlib==3.7.4
numpy==1.24.4
packaging==23.2
pillow==10.2.0
pycocotools @ git+https://github.com/philferriere/cocoapi.git@2929bd2ef6b451054755dfd7ceb09278f935f7ad#subdirectory=PythonAPI
pyparsing==3.1.1
python-dateutil==2.8.2
scipy==1.10.1
six==1.16.0
torch==1.10.1+cpu
torchaudio==0.10.1+cpu
torchvision==0.11.2+cpu
typing_extensions==4.9.0
zipp==3.17.0
